haproxy: Set initial state to backup in keepalived and care about a proper priority
Only the server with the highest priority should come up as master.
Having multiple servers coming up as master is not a good idea. But it is
okay to have all servers configured "backup" as their initial state - with
the impact that there is a delay of the server with the highest priority
becoming active. Note that in order for this to work properly the
priority of the server with the highest one, needs to be at least 50
higher than the server with the next highest priority.